requirements.txt
pyautogui
pyperclip
requests
PyYAML
mss
rapidocr-onnxruntime
numpy
flask
